+++ worklab-frontend/src/router.jsx
@@ -3,6 +3,7 @@ import { createBrowserRouter } from 'react-router-dom'
 import App from './App'
 import Pacientes from './pages/Pacientes'
 import Exames from './pages/Exames'
+import FormularioPaciente from './pages/FormularioPaciente'
 
 const router = createBrowserRouter([
   {
@@ -18,6 +19,14 @@ const router = createBrowserRouter([
         element: <Exames />
       },
     ]
+  },
+  {
+    path: 'cadastrar/paciente',
+    element: <FormularioPaciente />
+  },
+  {
+    path: 'cadastrar/exames',
+    element: ''
   }
 ])
